Add files via upload
If you are a person who loves to jot down the experience of his day this is a good place. The layout is simple with date, a box to write down your feelings and three buttons "share my thoughts" for sharing your thought with your friends,"Forget it" for clearing the box and "keep it" to download the thoughts locally. This project is made with Vite+Vue3 and the developer prefers calling it "I LOVE MY DIARY". 
API used :Web Share APi
Library used: moment library